+++ pmaw/PushshiftAPIBase.py
@@ -1,7 +1,6 @@
 import json
 import copy
 import logging
-import sys
 from concurrent.futures import ThreadPoolExecutor, as_completed
 
 import requests