+++ CHANGES.md
@@ -1,3 +1,11 @@
+## 3.0.0 (2022/12/2x)
+
+- changed `before` and `after` to `until` and `since`
+- removed `metadata=true` as this is now always enabled
+- set `order='desc'` as this replaces `sort`
+- set `sort='created_utc'` so that slicing still works as expected
+- Read more on [COLO switchover](https://www.reddit.com/r/pushshift/comments/zkggt0/update_on_colo_switchover_bug_fixes_reindexing/)
+
 ## 2.1.3 (2022/02/20)
 
 - Don't inherit from object in classes